+++ src/client/src/api-lib/index.js
@@ -81,7 +81,7 @@ import {
   getLayers,
 } from "./actions/stratigraphy";
 
-import { getProfile, getProfiles, getProfileLayers } from "./actions/profile";
+import { getProfileLayers } from "./actions/profile";
 
 import { loadDomains, patchCodeConfig } from "./actions/domains";
 
@@ -155,8 +155,6 @@ export {
   deleteLayer,
   gapLayer,
   getLayers,
-  getProfile,
-  getProfiles,
   getProfileLayers,
   loadDomains,
   patchCodeConfig,