+++ tests/Controllers/StratigraphyControllerTest.cs
@@ -71,10 +71,10 @@ public class StratigraphyControllerTest
 
         Assert.AreEqual(1000017, stratigraphy.BoreholeId);
         Assert.AreEqual(3000, stratigraphy.KindId);
-        Assert.AreEqual("Kailee Wisozk", stratigraphy.Name);
-        Assert.AreEqual("It only works when I'm Wake Island.", stratigraphy.Notes);
+        Assert.AreEqual("Marie Parker", stratigraphy.Name);
+        Assert.AreEqual("i use it for 10 weeks when i'm in my jail.", stratigraphy.Notes);
         Assert.AreEqual(1, stratigraphy.CreatedById);
-        Assert.AreEqual(3, stratigraphy.UpdatedById);
+        Assert.AreEqual(2, stratigraphy.UpdatedById);
         Assert.AreEqual(false, stratigraphy.IsPrimary);
     }
 
@@ -85,8 +85,8 @@ public class StratigraphyControllerTest
 
         var stratigraphy = ActionResultAssert.IsOkObjectResult<Stratigraphy>(stratigraphyResult.Result);
         Assert.AreEqual(1002423, stratigraphy.BoreholeId);
-        Assert.AreEqual("Virginia Ortiz", stratigraphy.Name);
-        Assert.AreEqual("I saw one of these in Grenada and I bought one.", stratigraphy.Notes);
+        Assert.AreEqual("Jaycee Beahan", stratigraphy.Name);
+        Assert.AreEqual("It only works when I'm Kuwait.", stratigraphy.Notes);
     }
 
     [TestMethod]
@@ -113,9 +113,9 @@ public class StratigraphyControllerTest
         Assert.IsInstanceOfType(copiedStratigraphyId, typeof(int));
         var copiedStratigraphy = GetStratigraphy((int)copiedStratigraphyId);
 
-        Assert.AreEqual("Virginia Ortiz (Clone)", copiedStratigraphy.Name);
+        Assert.AreEqual("Jaycee Beahan (Clone)", copiedStratigraphy.Name);
         Assert.AreEqual("sub_admin", copiedStratigraphy.CreatedBy.SubjectId);
-        Assert.AreEqual("sub_editor", copiedStratigraphy.UpdatedBy.SubjectId);
+        Assert.AreEqual("sub_controller", copiedStratigraphy.UpdatedBy.SubjectId);
         Assert.AreEqual(false, copiedStratigraphy.IsPrimary);
         Assert.AreSame(originalStratigraphy.Kind, copiedStratigraphy.Kind);
 
@@ -260,6 +260,7 @@ public class StratigraphyControllerTest
             BoreholeId = boreholeWithoutStratigraphy.Id,
             Name = "KODACLUSTER",
             Notes = "ARGONTITAN",
+            QualityId = 9003,
         };
 
         var createResult = await controller.CreateAsync(stratigraphyToAdd);
@@ -267,7 +268,7 @@ public class StratigraphyControllerTest
 
         var createdStratigraphy = (Stratigraphy?)((OkObjectResult)createResult.Result!).Value;
         createdStratigraphy = GetStratigraphy(createdStratigraphy.Id);
-        AssertStratigraphy(createdStratigraphy, boreholeWithoutStratigraphy.Id, "KODACLUSTER", "ARGONTITAN");
+        AssertStratigraphy(createdStratigraphy, boreholeWithoutStratigraphy.Id, "KODACLUSTER", "ARGONTITAN", 9003);
 
         // Because the stratigraphy is the first one for the borehole, it is automatically the primary stratigraphy.
         Assert.AreEqual(true, createdStratigraphy.IsPrimary);
@@ -405,6 +406,7 @@ public class StratigraphyControllerTest
         stratigraphyToEdit.Date = new DateTime(1999, 9, 9).ToUniversalTime();
         stratigraphyToEdit.Name = "ERRONEOUS";
         stratigraphyToEdit.Notes = "REDPOINT";
+        stratigraphyToEdit.QualityId = 9000;
 
         var editResult = await controller.EditAsync(stratigraphyToEdit);
         var editedStratigraphy = ActionResultAssert.IsOkObjectResult<Stratigraphy>(editResult.Result);
@@ -414,6 +416,7 @@ public class StratigraphyControllerTest
         Assert.AreEqual(new DateTime(1999, 9, 9).ToUniversalTime(), editedStratigraphy.Date);
         Assert.AreEqual("ERRONEOUS", editedStratigraphy.Name);
         Assert.AreEqual("REDPOINT", editedStratigraphy.Notes);
+        Assert.AreEqual(9000, editedStratigraphy.QualityId);
     }
 
     [TestMethod]
@@ -467,12 +470,13 @@ public class StratigraphyControllerTest
         ActionResultAssert.IsInternalServerError(editResult.Result, "locked");
     }
 
-    private void AssertStratigraphy(Stratigraphy actual, int expectedBoreholeId, string exptectedName, string expectedNotes)
+    private void AssertStratigraphy(Stratigraphy actual, int expectedBoreholeId, string exptectedName, string expectedNotes, int? qualityId = null)
     {
         Assert.AreEqual(StratigraphyController.StratigraphyKindId, actual.KindId);
         Assert.AreEqual(expectedBoreholeId, actual.BoreholeId);
         Assert.AreEqual(exptectedName, actual.Name);
         Assert.AreEqual(expectedNotes, actual.Notes);
+        Assert.AreEqual(qualityId, actual.QualityId);
     }
 
     private void SetupControllerWithAlwaysLockedBorehole()